requirements.txt
httpx
beautifulsoup4
lxml
loguru
tenacity
pydantic>=2.10
python-dotenv
playwright==1.55.0
fastapi
uvicorn[standard]
psycopg2-binary
pytesseract
pillow
crawl4ai
supabase
